requirements.txt
matplotlib>=3.4
pandas
requests
pillow
tqdm
python-dotenv 